Create a get_device() function
